make submission: division result type fix (python3 issue)

+++ pylearn2/scripts/icml_2013_wrepl/emotions/make_submission.py
@@ -65,7 +65,7 @@ f = function([X], y)
 
 y = []
 
-for i in xrange(dataset.X.shape[0] / batch_size):
+for i in xrange(int(dataset.X.shape[0] / batch_size)):
     x_arg = dataset.X[i*batch_size:(i+1)*batch_size,:]
     if X.ndim > 2:
         x_arg = dataset.get_topological_view(x_arg)
